requirements.txt
yfinance>=0.2.40
pandas>=2.2.2
numpy>=1.26.4
ta>=0.11.0
flask>=3.0.3
requests>=2.32.3
python-dotenv>=1.0.1
anthropic>=0.28.0
feedparser>=6.0.11